# api/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Rinna API",
    "description": "API for the Rinna project management system\n\n## Security\n\nThe Rinna API implements several security features to protect your data:\n\n- API Key Authentication: All endpoints require a valid API key in the Authorization header\n- Rate Limiting: Configurable request limits to prevent abuse\n- OAuth Integration: Connect to third-party services securely\n- Webhook Security: End-to-end verification of webhook payloads\n- Request Tracing: All requests include a unique X-Request-ID for traceability\n\n## Rate Limiting\n\nRate limiting is applied to all API endpoints to ensure fair usage and system stability. When rate limited, you will receive a 429 Too Many Requests response with a Retry-After header.\n\nRate limit headers:\n- X-RateLimit-Limit: Maximum requests allowed per minute\n- X-RateLimit-Remaining: Remaining requests in the current window\n- X-RateLimit-Reset: Timestamp when the limit resets (Unix time)\n\n## Detailed Documentation\n\nFor more detailed documentation, please refer to:\n- [API Security Guide](./docs/api-security-guide.md)\n- [Rate Limiting Documentation](./docs/rate-limiting.md)\n- [Secure Integration Examples](./docs/secure-integration-examples.md)\n",
    "version": "1.6.6",
    "contact": {
      "name": "Eric C. Mumford",
      "email": "info@example.com"
    },
    "license": {
      "name": "MIT"
    }
  },
  "host": "localhost:8080",
  "basePath": "/api/v1",
  "schemes": [
...
  "consumes": [
...
  "produces": [
...
  "securityDefinitions": {
    "ApiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "API key authentication using the Bearer scheme. Example: 'Authorization: Bearer ri-api-key'"
    },
    "OAuth2": {
      "type": "oauth2",
      "flow": "accessCode",
      "authorizationUrl": "https://localhost:8080/api/v1/oauth/authorize",
      "tokenUrl": "https://localhost:8080/api/v1/oauth/token",
      "scopes": {
        "read": "Read access to projects and work items",
        "write": "Write access to projects and work items",
        "admin": "Administrative access"
      }
    }
  },
  "paths": {
...
    "/oauth/authorize/{provider}": {
      "get": {
        "tags": [
          "oauth"
        ],
        "summary": "Get OAuth authorization URL",
        "description": "Generates an authorization URL for the specified OAuth provider.\nThis is the first step in the OAuth flow to connect to external services.\n",
        "operationId": "getOAuthAuthorizationURL",
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "description": "OAuth provider (github, jira, azure-devops, gitlab, bitbucket)",
            "required": true,
            "type": "string",
            "enum": [
              "github",
              "jira",
              "azure-devops",
              "gitlab",
              "bitbucket"
            ]
          },
          {
            "name": "project_id",
            "in": "query",
            "description": "Project ID to associate with the OAuth token",
            "required": true,
            "type": "string"
          },
          {
            "name": "user_id",
            "in": "query",
            "description": "User ID to associate with the OAuth token",
            "required": true,
            "type": "string"
          },
          {
            "name": "redirect_uri",
            "in": "query",
            "description": "URI to redirect to after authorization (must be whitelisted)",
            "required": false,
            "type": "string"
          }
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OAuth authorization URL generated successfully",
            "schema": {
              "type": "object",
              "properties": {
                "authorization_url": {
                  "type": "string",
                  "description": "URL to redirect the user to for OAuth authorization"
                },
                "state": {
                  "type": "string",
                  "description": "State token to validate the callback"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters"
          },
          "401": {
            "description": "Unauthorized"
          },
          "429": {
            "description": "Rate limit exceeded"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/oauth/callback": {
      "get": {
        "tags": [
          "oauth"
        ],
        "summary": "OAuth callback handler",
        "description": "Handles the callback from OAuth providers after user authorization.\nExchanges the authorization code for access tokens.\n",
        "operationId": "handleOAuthCallback",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "description": "Authorization code from OAuth provider",
            "required": true,
            "type": "string"
          },
          {
            "name": "state",
            "in": "query",
            "description": "State token to validate the request",
            "required": true,
            "type": "string"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OAuth authorization successful",
            "schema": {
              "$ref": "#/definitions/OAuthToken"
            }
          },
          "400": {
            "description": "Invalid request parameters"
          },
          "401": {
            "description": "Invalid state token"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/oauth/tokens": {
      "get": {
        "tags": [
          "oauth"
        ],
        "summary": "List OAuth tokens",
        "description": "Lists all OAuth tokens for a provider and project",
        "operationId": "listOAuthTokens",
        "parameters": [
          {
            "name": "provider",
            "in": "query",
            "description": "OAuth provider to filter by",
            "required": false,
            "type": "string",
            "enum": [
              "github",
              "jira",
              "azure-devops",
              "gitlab",
              "bitbucket"
            ]
          },
          {
            "name": "project_id",
            "in": "query",
            "description": "Project ID to filter by",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "List of OAuth tokens",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/OAuthToken"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "429": {
            "description": "Rate limit exceeded"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/oauth/tokens/{provider}": {
      "delete": {
        "tags": [
          "oauth"
        ],
        "summary": "Revoke OAuth token",
        "description": "Revokes and removes an OAuth token for a provider, project, and user",
        "operationId": "revokeOAuthToken",
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "description": "OAuth provider",
            "required": true,
            "type": "string",
            "enum": [
              "github",
              "jira",
              "azure-devops",
              "gitlab",
              "bitbucket"
            ]
          },
          {
            "name": "project_id",
            "in": "query",
            "description": "Project ID associated with the token",
            "required": true,
            "type": "string"
          },
          {
            "name": "user_id",
            "in": "query",
            "description": "User ID associated with the token",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "204": {
            "description": "OAuth token revoked successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Token not found"
          },
          "429": {
            "description": "Rate limited"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/oauth/providers": {
      "get": {
        "tags": [
          "oauth"
        ],
        "summary": "List available OAuth providers",
        "description": "Returns a list of available OAuth providers and their configuration requirements",
        "operationId": "listOAuthProviders",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "List of OAuth providers",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/OAuthProvider"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "429": {
            "description": "Rate limited"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "OAuthToken": {
      "type": "object",
      "properties": {
        "access_token": {
          "type": "string",
          "description": "The OAuth access token"
        },
        "token_type": {
          "type": "string",
          "description": "The token type (usually \"Bearer\")"
        },
        "refresh_token": {
          "type": "string",
          "description": "The refresh token (if available)"
        },
        "expiry": {
          "type": "string",
          "format": "date-time",
          "description": "When the token expires"
        },
        "scopes": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "description": "The scopes this token has access to"
        },
        "provider": {
          "type": "string",
          "description": "The OAuth provider (github, jira, etc.)"
        },
        "user_id": {
          "type": "string",
          "description": "The user ID associated with this token"
        },
        "project_id": {
          "type": "string",
          "description": "The project ID associated with this token"
        },
        "created_at": {
          "type": "string",
          "format": "date-time",
          "description": "When the token was created"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time",
          "description": "When the token was last updated"
        },
        "metadata": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "description": "Additional metadata about the token"
        }
      }
    },
    "OAuthProvider": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "description": "Provider identifier (e.g., github, jira)"
        },
        "name": {
          "type": "string",
          "description": "Display name of the provider"
        },
        "description": {
          "type": "string",
          "description": "Description of the provider"
        },
        "icon_url": {
          "type": "string",
          "description": "URL to the provider's icon"
        },
        "auth_url_template": {
          "type": "string",
          "description": "Template for constructing the authorization URL"
        },
        "token_url": {
          "type": "string",
          "description": "URL for token exchange"
        },
        "scopes": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "description": "Scope identifier"
              },
              "name": {
                "type": "string",
                "description": "Display name of the scope"
              },
              "description": {
                "type": "string",
                "description": "Description of what the scope grants access to"
              }
            }
          }
        },
        "configuration_fields": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "description": "Field name"
              },
              "label": {
                "type": "string",
                "description": "Display label for the field"
              },
              "required": {
                "type": "boolean",
                "description": "Whether the field is required"
              },
              "type": {
                "type": "string",
                "description": "Field type (string, boolean, number)"
              }
            }
          }
        },
        "documentation_url": {
          "type": "string",
          "description": "URL to documentation about integrating with this provider"
        }
      }
    },
...
  }
}